Refactor project module structure
Split the Orison project implementation into focused project modules, keep the package entrypoint as a facade, update package metadata and version to 2.2.0, and align the publish skill with the current release workflow.
