+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "orison",
-  "version": "2.1.0",
+  "version": "2.2.0",
   "description": "A static site generator built on top of lit-html.",
   "keywords": [
     "Static site generation",
@@ -12,7 +12,7 @@
   "homepage": "https://orison.alexlockhart.me",
   "repository": {
     "type": "git",
-    "url": "https://github.com/megazear7/orison"
+    "url": "git+https://github.com/megazear7/orison.git"
   },
   "engines": {
     "node": ">=24.16.0"
@@ -61,7 +61,7 @@
     "test": "run-s orison:build docs:build"
   },
   "bin": {
-    "orison": "./bin/cli.js"
+    "orison": "bin/cli.js"
   },
   "author": "Alex Lockhart",
   "license": "ISC"
